vote/requirements.txt
Flask
connexion[flask]
connexion[uvicorn]
httpx
PyYAML
gunicorn
